Order id not sent in queires bug fixed Order help screen order canceled close help screen

+++ app/src/main/java/com/tovo/eat/ui/account/orderhistory/ordersview/OrderHistoryActivityView.java
@@ -128,7 +128,7 @@ public class OrderHistoryActivityView extends BaseActivity<ActivityOrdersHistory
     @Override
     public void gotoSupport() {
         Intent intent = SupportActivity.newIntent(OrderHistoryActivityView.this);
-        intent.putExtra("orderid",Integer.parseInt(strOrderId));
+        intent.putExtra("orderid",Long.parseLong(strOrderId));
         intent.putExtra("type",AppConstants.QUERY_TYPE_ORDER_HISTORY);
         startActivity(intent);
     }
